Добавляет фильтрацию селектов
